Exclude kernels if they're in a higher priority value repo
DNF repo priorities are used to ensure that packages in a lower priority
value repo get chosen over packages in a higher priority value repo (yes,
that seems backwards, especially when trying to explain it).  The problem
is that dependencies override priorities, so if a kmod update in a lower
priority value repo requires a kernel in a higher priority value repo, that
kernel will be pulled in, even if there's a kernel in the lower priority
value repo that should be preferred.

This commit ensures that all kernels in higher priority value repos are
excluded, thus ensuring they can't accidentally be pulled in.  This does
cause issues with error messages to do broken kmod dependencies that will
be fixed in a future commit.

Signed-off-by: Jonathan Dieter <jdieter@ciq.com>