Exclude kmods without any matching kernels
When we exclude kernels due to repo priority, we may find ourselves with an
error message because there are updated kmods that depend on a kernel
that's no longer available.  In this case, DNF will give us a "none of the
providers can be installed" error, which is a negative experience for
plugin users.

This commit will now exclude any kmods without any matching kernels,
ensuring that DNF won't pop up any errors.

Signed-off-by: Jonathan Dieter <jdieter@ciq.com>