Code cleanup
This is mainly restructuring the code to ensure that we're not doing a
bunch of string/list manipulation unless we're in CLI mode.  This also
fixes some minor reporting issues around kmods and excluded kernels.

This commit also removes the code handling the old `kmod-kernel.d` path,
which was deprecated before this plugin was released.  Nothing should be
using that path anymore.

Signed-off-by: Jonathan Dieter <jdieter@ciq.com>