Temporarily disable plugin when available package sack isn't populated
We were seeing a warning:
```
WARNING: No kmod-idpf-irdma packages available in the repositories, so not
blocking updates based on kmod-idpf-irdma.
```
when `dnf list installed` was run.

This happened because `dnf list installed` doesn't populate the available
package cache, and the plugin assumed that kmod-idpf-irdma just wasn't
available.

To fix this, we're now just temporarily disabling the plugin if the
available package sack isn't populated.  This should only happen if `dnf
list installed` (or something similar) is run, which is irrelevant for the
plugin.

Signed-off-by: Jonathan Dieter <jdieter@ciq.com>